Add test case to reproduce crash on assigning reviews
